+++ src/lib/server/api/v1/auth_api.js
@@ -61,3 +61,40 @@ export async function logout(fetch) {
 	console.error('Logout failed');
 	await responseError(response);
 }
+
+/**
+ * Fetches the list of users from the server
+ * @param {typeof fetch} fetch
+ * @returns {Promise<*>}
+ */
+export async function listUsers(fetch) {
+	const response = await fetch(FRACTAL_SERVER_HOST + '/auth/userlist/', {
+		method: 'GET',
+		credentials: 'include'
+	});
+
+	if (response.ok) {
+		return await response.json();
+	}
+
+	await responseError(response);
+}
+
+/**
+ * Fetches a user from the server
+ * @param {typeof fetch} fetch
+ * @param {number|string} userId
+ * @returns {Promise<*>}
+ */
+export async function getUser(fetch, userId) {
+	const response = await fetch(`${FRACTAL_SERVER_HOST}/auth/users/${userId}/`, {
+		method: 'GET',
+		credentials: 'include'
+	});
+
+	if (response.ok) {
+		return await response.json();
+	}
+
+	await responseError(response);
+}