+++ src/routes/admin/+page.svelte
@@ -1,0 +1,6 @@
+<h3>Admin area</h3>
+
+<a href="/admin/users" class="btn btn-primary mt-3">
+	<i class="bi bi-people-fill" />
+	Manage users
+</a>