+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 This release requires fractal-server 1.4.0.
 
+* Added user profile page (\#336).
+* Added admin area with users management (\#336).
 * Added Jobs button in home page (\#346).
 * Improved jobs table layout for small screens (\#346).
 * Fixed jobs status badge color bug (\#346).